Avoid C-style cast in base

+++ include/deal.II/base/point.h
@@ -372,7 +372,7 @@ template <int dim, typename Number>
 inline Number
 Point<dim, Number>::operator()(const unsigned int index) const
 {
-  AssertIndexRange((int)index, dim);
+  AssertIndexRange(index, dim);
   return this->values[index];
 }
 
@@ -382,7 +382,7 @@ template <int dim, typename Number>
 inline Number &
 Point<dim, Number>::operator()(const unsigned int index)
 {
-  AssertIndexRange((int)index, dim);
+  AssertIndexRange(index, dim);
   return this->values[index];
 }
